Extract task-queue framework to parent package
